240818/requirements.txt
streamlit 
plotly.express 
pandas 
numpy